Final version
Changed the interview scenes to stage. Made the audience, stage, and the speaker. If the mouse is clicked, the audience's eyes will change to red, spotlight shines on the speaker, and the speaker starts to sweat.
Because of that the next phase is the heartbeat where it shows that the speaker heartbeat is increasing, or the heart is beating faster, so I put the heartbeat screen